[Moviearchiver] completely overworked
- code was summarized in one file
- color button labels have been reactivated
- skins have been reduced to HD, they are automatically upscaled to fHD
- all GLOBAL variables have been moved to the MAglobs class
- the German translations have been adapted accordingly

HINT: **MovieArchiver** is still working under Python2 and Python3
