--- src/Makefile.am
@@ -1,3 +1,0 @@
-installdir = $(libdir)/enigma2/python/Plugins/Extensions/MovieArchiver
-install_PYTHON = *.py
-install_DATA = plugin.png maintainer.info
