+++ src/__init__.py
@@ -1,4 +1,3 @@
-# -*- coding: UTF-8 -*-
 #######################################################################
 #
 #    MovieArchiver
@@ -21,14 +20,13 @@
 #
 #######################################################################
 
-from Components.config import config, ConfigSubsection, ConfigNumber, ConfigText, ConfigYesNo, ConfigLocations
-from Tools.Directories import resolveFilename, SCOPE_HDD, SCOPE_PLUGINS
-from Components.Language import language
+# PYTHON IMPORTS
 from gettext import bindtextdomain, dgettext, gettext
 
-#############################################################
-
-# Gettext
+# ENIGMA IMPORTS
+from Components.config import config, ConfigSubsection, ConfigNumber, ConfigText, ConfigYesNo, ConfigLocations
+from Components.Language import language
+from Tools.Directories import resolveFilename, SCOPE_HDD, SCOPE_PLUGINS
 
 PluginLanguageDomain = "MovieArchiver"
 PluginLanguagePath = "Extensions/MovieArchiver/locale"
@@ -39,21 +37,15 @@ def localeInit():
 
 
 def _(txt):
-	if dgettext(PluginLanguageDomain, txt):
-		return dgettext(PluginLanguageDomain, txt)
-	else:
-		return gettext(txt)
+	return dgettext(PluginLanguageDomain, txt) if dgettext(PluginLanguageDomain, txt) else gettext(txt)
+
 
 localeInit()
 language.addCallback(localeInit)
 
-#############################################################
-
 
 def printToConsole(msg):
 	print("[MovieArchiver] %s" % msg)
-
-#############################################################
 
 
 # Define Settings Entries
@@ -62,28 +54,17 @@ config.plugins.MovieArchiver.enabled = ConfigYesNo(default=False)
 config.plugins.MovieArchiver.backup = ConfigYesNo(default=False)
 config.plugins.MovieArchiver.skipDuringRecords = ConfigYesNo(default=True)
 config.plugins.MovieArchiver.showLimitReachedNotification = ConfigYesNo(default=True)
-
-# default hdd
-defaultDir = resolveFilename(SCOPE_HDD)
+defaultDir = resolveFilename(SCOPE_HDD)  # default hdd
 if config.movielist.videodirs.getValue() and len(config.movielist.videodirs.getValue()) > 0:
 	defaultDir = config.movielist.videodirs.getValue()[0]
-
 config.plugins.MovieArchiver.sourcePath = ConfigText(default=defaultDir, fixed_size=False, visible_width=30)
 config.plugins.MovieArchiver.sourcePath.lastValue = config.plugins.MovieArchiver.sourcePath.getValue()
-
 config.plugins.MovieArchiver.sourceLimit = ConfigNumber(default=30)
-
-
-# exclude folders
-config.plugins.MovieArchiver.excludeDirs = ConfigLocations(visible_width=30)
-
+config.plugins.MovieArchiver.excludeDirs = ConfigLocations(visible_width=30)  # exclude folders
 config.plugins.MovieArchiver.targetPath = ConfigText(default=defaultDir, fixed_size=False, visible_width=30)
 config.plugins.MovieArchiver.targetPath.lastValue = config.plugins.MovieArchiver.targetPath.getValue()
+config.plugins.MovieArchiver.targetLimit = ConfigNumber(default=30)  # interval
 
-# interval
-config.plugins.MovieArchiver.targetLimit = ConfigNumber(default=30)
-
-#############################################################
 # Helper Functions
 
 
@@ -101,8 +82,6 @@ def getTargetPath():
 
 def getTargetPathValue():
 	return getTargetPath().getValue()
-
-#############################################################
 
 
 __all__ = ['_', 'config', 'printToConsole', 'getSourcePath', 'getSourcePathValue', 'getTargetPath', 'getTargetPathValue']
