--- src/maintainer.info
@@ -1,2 +1,0 @@
-svox
-MovieArchiver
